moving + kinesis + refactor
