+++ tests/unit/parser/test_apigwv2.py
@@ -5,8 +5,8 @@ from aws_lambda_powertools.utilities.parser.models import (
     RequestContextV2Authorizer,
 )
 from aws_lambda_powertools.utilities.typing import LambdaContext
-from tests.functional.parser.schemas import MyApiGatewayBusiness
 from tests.functional.utils import load_event
+from tests.unit.parser.schemas import MyApiGatewayBusiness
 
 
 @event_parser(model=MyApiGatewayBusiness, envelope=envelopes.ApiGatewayV2Envelope)
