+++ tests/unit/parser/test_dynamodb.py
@@ -8,8 +8,8 @@ from aws_lambda_powertools.utilities.parser import (
     event_parser,
 )
 from aws_lambda_powertools.utilities.typing import LambdaContext
-from tests.functional.parser.schemas import MyAdvancedDynamoBusiness, MyDynamoBusiness
 from tests.functional.utils import load_event
+from tests.unit.parser.schemas import MyAdvancedDynamoBusiness, MyDynamoBusiness
 
 
 @event_parser(model=MyDynamoBusiness, envelope=envelopes.DynamoDBStreamEnvelope)
