+++ tests/unit/parser/test_eventbridge.py
@@ -6,11 +6,11 @@ from aws_lambda_powertools.utilities.parser import (
     event_parser,
 )
 from aws_lambda_powertools.utilities.typing import LambdaContext
-from tests.functional.parser.schemas import (
+from tests.functional.utils import load_event
+from tests.unit.parser.schemas import (
     MyAdvancedEventbridgeBusiness,
     MyEventbridgeBusiness,
 )
-from tests.functional.utils import load_event
 
 
 @event_parser(model=MyEventbridgeBusiness, envelope=envelopes.EventBridgeEnvelope)
